+++ tests/testthat/test_Read_and_Split_VCFs.R
@@ -144,6 +144,9 @@ test_that("Test ReadAndSplitVCFs always.merge.SBS", {
 })
 
 test_that("Test ReadAndSplitVCFs filter.status argument", {
+  skip_if("" == system.file(package = "BSgenome.Hsapiens.1000genomes.hs37d5"))
+  stopifnot(requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5"))
+  
   file1 <- "testdata/Strelka-SBS-GRCh37/Strelka.SBS.GRCh37.s6.vcf"
   vcf1 <- ReadVCF(file = file1, filter.status = "PASS")
   expect_equal(nrow(vcf1), 3)
@@ -191,6 +194,9 @@ test_that("Test ReadAndSplitVCFs filter.status argument", {
 })
 
 test_that("Test ReadAndSplitVCFs dealing with VCF with no 'FILTER' column", {
+  skip_if("" == system.file(package = "BSgenome.Hsapiens.1000genomes.hs37d5"))
+  stopifnot(requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5"))
+  
   file1 <- "testdata/Strelka.SBS.GRCh38.no.FILTER.column.vcf"
   vcf1 <- expect_warning(ReadVCF(file = file1, filter.status = "PASS"))
   expect_equal(nrow(vcf1), 7)